mysite/requirements.txt
django-haystack==3.0
whoosh
jieba
